requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
backtrader
matplotlib
pandas-ta
TA-Lib
scipy
plotly
transformers
torch
tensorflow
alpaca-trade-api
sqlmodel
SQLAlchemy
psycopg[binary]
passlib[bcrypt]
PyJWT
polygon-api-client
email-validator
